Keep track of iterators and close them before closing the DB
Before closing the RocksDB database, first all open iterators
are closed before closing the main database to avoid an
inconsistent state with invalid databaset iterators.

Related to issue #113.
