+++ src/test/java/focusedCrawler/link/frontier/FrontierTest.java
@@ -20,20 +20,20 @@ import org.junit.rules.TemporaryFolder;
 import focusedCrawler.util.persistence.PersistentHashtable.DB;
 
 public class FrontierTest {
-    
+
     @Rule
     public TemporaryFolder tempFolder = new TemporaryFolder();
 
     public Path testPath;
-    
+
     private Frontier frontier;
-    
+
     @Before
     public void setUp() throws IOException {
         testPath = Paths.get(tempFolder.newFolder().toString());
         frontier = new Frontier(testPath.toString(), 1000, DB.ROCKSDB);
     }
-    
+
     @After
     public void tearDown() throws IOException {
         frontier.close();
@@ -45,44 +45,43 @@ public class FrontierTest {
         // given
         LinkRelevance link1 = new LinkRelevance(new URL("http://www.example1.com/index.html"), 1);
         LinkRelevance link2 = new LinkRelevance(new URL("http://www.example2.com/index.html"), 1);
-        
+
         // when
         frontier.insert(link1);
-        
+
         // then
         assertThat(frontier.exist(link1), is(1d));
         assertThat(frontier.exist(link2), is(nullValue()));
     }
-    
+
     @Test
     public void shouldInsertUrlsAndSelectGivenNumberOfUrls() throws Exception {
         // given
         LinkRelevance link1 = new LinkRelevance(new URL("http://www.example1.com/index.html"), 1);
         LinkRelevance link2 = new LinkRelevance(new URL("http://www.example2.com/index.html"), 2);
-        
+
         // when
         frontier.insert(link1);
         frontier.insert(link2);
-        
+
         // then
         assertThat(frontier.exist(link1), is(notNullValue()));
         assertThat(frontier.exist(link1), is(1d));
-        
+
         assertThat(frontier.exist(link2), is(notNullValue()));
         assertThat(frontier.exist(link2), is(2d));
     }
-    
-    
+
     @Test
     public void shouldInsertAndDeleteUrl() throws Exception {
         // given
         LinkRelevance link1 = new LinkRelevance(new URL("http://www.example1.com/index.html"), 1);
-        
+
         // when
         frontier.insert(link1);
         // then
         assertThat(frontier.exist(link1), is(1d));
-        
+
         // when
         frontier.delete(link1);
         // then
